requirements.txt
Flask==2.2.2
nltk==3.8.0
textblob==0.18.0.post0
newspaper3k==0.2.8